btracker/requirements.txt
alembic~=0.9.9
connexion~=2.3.0
datadog~=0.22.0
flake8~=3.5.0
Flask-Cors~=3.0.6
freezegun~=0.3.10
gevent~=1.2.2
lxml~=4.2.3
moto~=1.3.3
pip-tools~=2.0.2
psycogreen~=1.0
psycopg2~=2.7.5
pytest-ordering~=0.5
pytest-cov~=2.5.1
pytest-watch~=4.2.0
python-dateutil~=2.5.3
pytz~=2018.3
simplejson~=3.14.0
SQLAlchemy~=1.2.7
